refactor: minimized client project indentation and simplifed project structure.
