+++ clients/cli/src/bundle/mod.rs
@@ -1,4 +1,1 @@
-mod download;
 pub mod epub;
-
-pub use download::{download, DownloadOptions};
