+++ clients/cli/src/data/mod.rs
@@ -1,0 +1,5 @@
+mod global_tracking;
+mod novel_tracking;
+
+pub use global_tracking::{GlobalData, GlobalTracker};
+pub use novel_tracking::{DownloadLog, EventKind, NovelTracking};
